chore(js): add e2e test to check built package.json
